🛠️refactor: 크루별 출석 히스토리 dto 구현 및 출석 히스토리 출력

+++ src/test/java/domain/AttendanceBookTest.java
@@ -117,7 +117,7 @@ public class AttendanceBookTest {
         final String name = "쿠키";
 
         //when
-        final Map<AttendanceStatus, Long> countAttendanceStatus = attendanceBook.countAttendanceStatus(name);
+        final Map<AttendanceStatus, Integer> countAttendanceStatus = attendanceBook.countAttendanceStatus(name);
 
         //then
         assertAll(
